add lca

+++ master.cpp
@@ -3,7 +3,7 @@
 using ll = long long;
 using namespace std;
 const int INFint = 1e9+1;
-const ll INFll = 1e18+1;
+const ll INFll = (ll)1e18+1;
 ll MOD=1e9+7;
